requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
urllib3>=2.0.0
